requirements.txt
python-magic
python-magic-bin
python-magic-win64